Micro-optimizations in XmlNode
- The xpath delimiter is now just a char
- No need to copy strings which are not modified
- `find()` + relevant calls to `substr()` will avoid searching for the split
  position three times
- The test regarding whether the xpath is made of several parts has be moved
  out of the loop
- A couple of getters have been made inline
- The const version of `findFirstNode()` has been changed to return a const
  reference instead of a value -- when nothing is found, a reference to an
  internal global null node is returned
- A few comparisons to `""` have be replaced with a check of `empty()`

In the end `(const) XmlNode::findFirstNode()` takes less than half the cycles
which it used to take -- The cycles were measured with callgrind on a program
compiled with `-O2 -g -DNDEBUG`.

More improvments could be made on the xml document construction by introducing
something like `std::experimental::string_view` in order to avoid string
duplication as much as possible.