Support M3 compressed data.
M3 data is the same as C3 (jpeg compression) with the addition of
a block mask.  This block mask indicates which blocks are empty
and not included in the file.

Remove faulty assumption that M3 data must have a pad pixel mask
even if the header says the pad pixel mask does not exist.

Remap the block offsets/sizes from masked to absolute so the normal
block lookup procedure works.

Tested on the AFRL Greene 07 dataset.